requirements.txt
Django~=5.0.2
requests~=2.31.0
geopy~=2.4.1
djangorestframework~=3.14.0